Re-organizing Leaflet layer views
