+++ src/geo/leaflet-extensions.js
@@ -1,7 +1,9 @@
-var L = {};
-L.CartoDBGroupLayerBase = require('./leaflet/leaflet-cartodb-group-layer-base');
-L.CartoDBGroupLayer = require('./leaflet/leaflet-cartodb-group-layer');
-L.NamedMap = require('./leaflet/leaflet-named-map');
-L.CartoDBLayer = require('./leaflet/leaflet-cartodb-layer');
+// var L = {};
 
-module.exports = L;
+// // TODO: Unify this into one single layer
+// L.CartoDBGroupLayerBase = require('./leaflet/leaflet-cartodb-group-layer-base');
+// L.CartoDBGroupLayer = require('./leaflet/leaflet-cartodb-group-layer');
+// L.NamedMap = require('./leaflet/leaflet-named-map');
+// L.CartoDBLayer = require('./leaflet/leaflet-cartodb-layer');
+
+// module.exports = L;
