--- src/geo/leaflet/leaflet-cartodb-group-layer.js
@@ -1,7 +1,0 @@
-var LeafletCartoDBGroupLayerBase = require('./leaflet-cartodb-group-layer-base');
-
-var LeafletCartoDBGroupLayer = LeafletCartoDBGroupLayerBase.extend({
-  _modelUpdated: function() {}
-});
-
-module.exports = LeafletCartoDBGroupLayer;
