--- src/geo/leaflet/leaflet-cartodb-named-map-view.js
@@ -1,6 +1,0 @@
-var layerView = require('./leaflet-layer-view-fn');
-var LeafletNamedMap = require('./leaflet-named-map');
-
-var LeafletCartoDBNamedMapView = layerView(LeafletNamedMap);
-
-module.exports = LeafletCartoDBNamedMapView;
