+++ src/geo/leaflet/leaflet-map-view.js
@@ -11,7 +11,6 @@ var LeafletLayerCartoDBView = require('./leaflet-layer-cartodb-view');
 var LeafletPlainLayerView = require('./leaflet-plain-layer-view');
 var LeafletGmapsTiledLayerView = require('./leaflet-gmaps-tiled-layer-view');
 var LeafletCartoDBLayerGroupView = require('./leaflet-cartodb-layer-group-view');
-var LeafletCartoDBNamedMapView = require('./leaflet-cartodb-named-map-view');
 var LeafletPointView = require('./leaflet-point-view');
 var LeafletPathView = require('./leaflet-path-view');
 
@@ -357,7 +356,7 @@ var LeafletMapView = MapView.extend({
     "gmapsbase": LeafletGmapsTiledLayerView,
 
     "layergroup": LeafletCartoDBLayerGroupView,
-    "namedmap": LeafletCartoDBNamedMapView,
+    "namedmap": LeafletCartoDBLayerGroupView,
     "torque": function(layer, map) {
       // TODO for now adding this error to be thrown if object is not present, since it's dependency
       // is not included in the standard bundle
